Feat: Diagnose error action component (#325)
* init diagnose-error component

* rm old diagnose error genui v1 component

* diagnose-error schema to shared/tools

* CR

* Update apps/web/src/components/thread/messages/ai.tsx

---------
