+++ apps/web/src/components/gen-ui/index.tsx
@@ -1,5 +1,4 @@
 import { ActionStep } from "./action-step";
-import { DiagnoseError } from "./diagnose-error";
 import { InitializeStep } from "./initialize-step";
 import { PushChanges } from "./push-changes";
 import { ReplanningStep } from "./replanning-step";
@@ -7,7 +6,6 @@ import { TaskSummary } from "./task-summary";
 
 export default {
   "action-step": ActionStep,
-  "diagnose-error": DiagnoseError,
   "initialize-step": InitializeStep,
   "push-changes": PushChanges,
   "replanning-step": ReplanningStep,
