+++ apps/web/src/components/v2/thread-view.tsx
@@ -127,7 +127,6 @@ export function ThreadView({
           setChatInput={setChatInput}
           handleSendMessage={handleSendMessage}
         />
-
         {/* Right Side - Actions & Plan */}
         <div className="flex h-full flex-1 flex-col">
           <div className="relative flex-1">
